Split out triple class into 3 single classes
